+++ src/render/shape/rect.rs
@@ -20,7 +20,7 @@ impl Rect {
 }
 
 impl Inset for Rect {
-    fn inset(mut self, amount: i32) -> Self {
+    fn inset(mut self, amount: i16) -> Self {
         self.size.width = self.size.width.saturating_add_signed(-2 * amount);
         self.size.height = self.size.height.saturating_add_signed(-2 * amount);
         self.origin.x += amount;
